refactor: 회원가입 로직 userController 모듈로 분리
- user.js에서 userController 모듈 불러오도록 변경
